modified:   c++/exam_text_invader_basic/exam_text_invader_basic.VC.db  	new file:   c++/exam_text_invader_basic/exam_text_invader_basic/ActorBulletCurve.cpp  	new file:   c++/exam_text_invader_basic/exam_text_invader_basic/ActorBulletCurve.h  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/Bullet.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/Bullet.h  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/EnemyFactorySample0.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/EnemyFactorySample1.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/EnemyFactorySample2.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/EnemyNormal.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/EnemyNormal.h  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/GameManager.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/GameManager.h  	new file:   c++/exam_text_invader_basic/exam_text_invader_basic/Score.cpp  	new file:   c++/exam_text_invader_basic/exam_text_invader_basic/Score.h  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/Timer.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/Timer.h  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/Unit.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/Unit.h  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/exam_text_invader_basic.cpp  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/exam_text_invader_basic.vcxproj  	modified:   c++/exam_text_invader_basic/exam_text_invader_basic/exam_text_invader_basic.vcxproj.filters  	new file:   c++/report/20170104_text_invader_ppt.txt  	new file:   c++/report/20170105_text_invader_report.txt

+++ c++/exam_text_invader_basic/exam_text_invader_basic/Timer.cpp
@@ -11,7 +11,7 @@ CTimer::~CTimer()
 {
 }
 
-void CTimer::SetInterval(float tInterval)
+void CTimer::SetInterval(int tInterval)
 {
 	mInterval = tInterval;
 }
